ISSUE-76: Moved all constants values
