+++ app/src/main/java/ar/com/tzulberti/archerytraining/database/consts/PlayoffConsts.java
@@ -1,4 +1,4 @@
-package ar.com.tzulberti.archerytraining.consts;
+package ar.com.tzulberti.archerytraining.database.consts;
 
 /**
  * Created by tzulberti on 6/2/17.
