+++ app/src/main/java/ar/com/tzulberti/archerytraining/database/consts/TournamentSerieConsts.java
@@ -1,4 +1,4 @@
-package ar.com.tzulberti.archerytraining.consts;
+package ar.com.tzulberti.archerytraining.database.consts;
 
 /**
  * Created by tzulberti on 5/16/17.
